# Avoid function call overhead during loops