Parseable was using a pattern whereby a listing function was called in nested loops. Removing this function call and replacing it with a single assignment to a variable increases performance of these functions by ~10x (clock time).

May have a positive impact on https://github.com/richardlehane/siegfried/issues/148

----

NB. just using the timing options available, i.e. unit test time, and wall clock. Siegfried unit tests are included as a convenience, it is expected this change is mostly for roy.

## Results

### Roy unit tests

#### Before
```
ross-spencer@exponential-decay:/tmp/sf/siegfried/cmd/roy$ go test
PASS
ok  	github.com/richardlehane/siegfried/cmd/roy	89.432s
```

#### After
```
ross-spencer@exponential-decay:/tmp/sf/siegfried/cmd/roy$ go test
PASS
ok  	github.com/richardlehane/siegfried/cmd/roy	9.342s
```

### SF unit tests

#### Before
```
ross-spencer@exponential-decay:/tmp/sf/siegfried/cmd/sf$ go test
PASS
ok  	github.com/richardlehane/siegfried/cmd/sf	7.616s
```

#### After
```
ross-spencer@exponential-decay:/tmp/sf/siegfried/cmd/sf$ go test
PASS
ok  	github.com/richardlehane/siegfried/cmd/sf	3.856s
```

### Extend signature

#### Before
```
ross-spencer@exponential-decay:/tmp/sf/siegfried/cmd/roy$ time ./roy build -extend <sigfile.xml>

real	0m6.493s
user	0m7.324s
sys	0m0.086s
```

#### After
```
ross-spencer@exponential-decay:/tmp/sf/siegfried/cmd/roy$ time ./roy build -extend <sigfile.xml>

real	0m0.284s
user	0m0.939s
sys	0m0.076s
```