@richardlehane 2025-09-18
thanks @ross-spencer this is awesome 
![SoExcited~GIF](https://github.com/user-attachments/assets/63d1e909-4ba0-4c6b-8d1b-4302f9f54020)